backend/requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pydantic
python-dotenv
requests
python-dateutil
